RF: widen CountBin to double per maintainer direction
Replaces the prior Design D float-bin commit per RAMitchell's review
direction on #8132 to use double rather than float for overflow safety
on large weighted counts. CountBin::x widens from int to double; matching
double accumulator promotion in Gini/Entropy GainPerSplit and SetLeafVector.
No CUB fast-path tricks; the unified BlockScan<BinT> path handles both
CountBin and AggregateBin.

Includes two adjacent cleanups carried forward from the prior PR-1 commit:
a static_cast<IdxT> in the ObjectiveTest fixture for the int-to-IdxT
narrowing the bin-type change surfaces, and removal of a stale
xfail_skl2onnx_bug on RandomForestClassifier in test_onnx.py.

See PR description for the bench numbers and the int64-vs-double
comparison RAMitchell requested.

+++ cpp/src/decisiontree/batched-levelalgo/bins.cuh
@@ -9,15 +9,17 @@ namespace ML {
 namespace DT {
 
 struct CountBin {
-  int x;
+  // double covers both the unweighted count path and the future weighted-count
+  // path with one bin type; 32-bit int would overflow on large weighted counts.
+  double x;
   CountBin(CountBin const&) = default;
-  HDI CountBin(int x_) : x(x_) {}
-  HDI CountBin() : x(0) {}
+  HDI CountBin(double x_) : x(x_) {}
+  HDI CountBin() : x(0.0) {}
 
   DI static void IncrementHistogram(CountBin* hist, int n_bins, int b, int label)
   {
     auto offset = label * n_bins + b;
-    CountBin::AtomicAdd(hist + offset, {1});
+    CountBin::AtomicAdd(hist + offset, {1.0});
   }
   DI static void AtomicAdd(CountBin* address, CountBin val) { atomicAdd(&address->x, val.x); }
   HDI CountBin& operator+=(const CountBin& b)
